fix: move user mysql tests into user/test/mysql
