+++ lib/user/test/index.js
@@ -1,15 +1,11 @@
 import assert from 'node:assert/strict'
-import crypto from 'node:crypto'
 import { describe, it, after, before } from 'node:test'
 
 import User from '../index.js'
 import Group from '../../group/index.js'
-import Mysql from '../../mysql.js'
 
 import userJson from './user.json' with { type: 'json' }
 import groupJson from '../../group/test/group.json' with { type: 'json' }
-
-const storeType = process.env.NICTOOL_DATA_STORE ?? 'mysql'
 
 // This suite soft-deletes and restores its user, so it uses private fixtures
 // rather than the shared 4096 user/group that the concurrently-run permission
@@ -168,132 +164,6 @@ describe('user', function () {
         password: userCase.password,
       })
       assert.ok(u)
-    })
-  })
-
-  // these drive the store through raw SQL, so they only apply to the MySQL repo.
-  // Left running under TOML they hang: the pool never connects and its open
-  // handle keeps the test process alive forever.
-  describe('password upgrade on login', { skip: storeType !== 'mysql' }, () => {
-    const upgradeUserId = 4200
-    const upgradeUser = {
-      nt_user_id: upgradeUserId,
-      nt_group_id: groupCase.id,
-      username: 'upgrade-test',
-      email: 'upgrade-test@example.com',
-      first_name: 'Upgrade',
-      last_name: 'Test',
-    }
-    const testPass = 'UpgradeMe!123'
-    const authCreds = {
-      username: `${upgradeUser.username}@${groupCase.name}`,
-      password: testPass,
-    }
-
-    async function getDbPassword() {
-      const rows = await Mysql.execute('SELECT password, pass_salt FROM nt_user WHERE nt_user_id = ?', [
-        upgradeUserId,
-      ])
-      return rows[0]
-    }
-
-    // Raw SQL so we can insert specific legacy password formats (plain text,
-    // SHA-1, PBKDF2-5000) that User.create() would hash on the way in.
-    async function insertUser(password, passSalt) {
-      await Mysql.execute(
-        'INSERT INTO nt_user (nt_user_id, nt_group_id, username, email, first_name, last_name, password, pass_salt) VALUES (?, ?, ?, ?, ?, ?, ?, ?)',
-        [
-          upgradeUserId,
-          upgradeUser.nt_group_id,
-          upgradeUser.username,
-          upgradeUser.email,
-          upgradeUser.first_name,
-          upgradeUser.last_name,
-          password,
-          passSalt,
-        ],
-      )
-    }
-
-    async function cleanup() {
-      await Mysql.execute('DELETE FROM nt_user WHERE nt_user_id = ?', [upgradeUserId])
-    }
-
-    before(cleanup)
-    after(cleanup)
-
-    it('upgrades plain text password to self-describing PBKDF2 on login', async () => {
-      await cleanup()
-      await insertUser(testPass, null)
-
-      const result = await User.authenticate(authCreds)
-      assert.ok(result, 'login should succeed')
-
-      const row = await getDbPassword()
-      assert.ok(row.pass_salt, 'pass_salt should be set after upgrade')
-      assert.notEqual(row.password, testPass, 'password should be hashed')
-      assert.ok(row.password.includes('$'), 'password should be in self-describing format')
-
-      // verify round-trip: can still log in with the upgraded hash
-      const again = await User.authenticate(authCreds)
-      assert.ok(again, 'login should succeed after upgrade')
-      await cleanup()
-    })
-
-    it('upgrades SHA1 password to self-describing PBKDF2 on login', async () => {
-      // authenticate() passes the full authTry.username (including @group) to
-      // validPassword(), so the HMAC key must match that full string
-      const sha1Hash = crypto
-        .createHmac('sha1', authCreds.username.toLowerCase())
-        .update(testPass)
-        .digest('hex')
-      await cleanup()
-      await insertUser(sha1Hash, null)
-
-      const result = await User.authenticate(authCreds)
-      assert.ok(result, 'login should succeed with SHA1 hash')
-
-      const row = await getDbPassword()
-      assert.ok(row.pass_salt, 'pass_salt should be set after upgrade')
-      assert.notEqual(row.password, sha1Hash, 'password should be re-hashed')
-      assert.ok(row.password.includes('$'), 'password should be in self-describing format')
-
-      const again = await User.authenticate(authCreds)
-      assert.ok(again, 'login should succeed after upgrade')
-      await cleanup()
-    })
-
-    it('upgrades PBKDF2-5000 to self-describing format on login', async () => {
-      const legacySalt = User.generateSalt()
-      const legacyHash = await User.hashAuthPbkdf2(testPass, legacySalt, 5000)
-      await cleanup()
-      await insertUser(legacyHash, legacySalt)
-
-      const result = await User.authenticate(authCreds)
-      assert.ok(result, 'login should succeed with legacy PBKDF2')
-
-      const row = await getDbPassword()
-      assert.notEqual(row.password, legacyHash, 'password should be re-hashed')
-      assert.notEqual(row.pass_salt, legacySalt, 'salt should be regenerated')
-      assert.ok(row.password.includes('$'), 'password should be in self-describing format')
-
-      const again = await User.authenticate(authCreds)
-      assert.ok(again, 'login should succeed after upgrade')
-      await cleanup()
-    })
-
-    it('does not re-hash password already in self-describing format', async () => {
-      const salt = User.generateSalt()
-      const hash = await User.hashForStorage(testPass, salt)
-      await cleanup()
-      await insertUser(hash, salt)
-
-      await User.authenticate(authCreds)
-
-      const row = await getDbPassword()
-      assert.equal(row.password, hash, 'password should be unchanged')
-      assert.equal(row.pass_salt, salt, 'salt should be unchanged')
-      await cleanup()
     })
   })
 })
